Working out loader file'
